// src/generated/models/request-stage-configuration-model.ts
import type { AccessRole } from './access-role-model';

export const RequestStageConfigurationStageTypeKeyToLabel = {
  'Provisioning': 'Provisioning',
  'GCCSync': 'GCC Sync',
  'AccessConfirmation': 'Access Confirmation',
  'NotificationReady': 'Notification Ready',
  'NotificationSent': 'Notification Sent'
} as const;
export type RequestStageConfigurationStageTypeKey = keyof typeof RequestStageConfigurationStageTypeKeyToLabel;

export interface RequestStageConfiguration {
  /**
   * @displayName Request Stage Configuration
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Request Stage Configuration Name
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  requestStageConfigurationName: string;
  /**
   * @displayName Access Role
   * @validationRule Required for create/update operations
   */
  accessRole: Pick<AccessRole, 'id' | 'accessRoleName'>;
  /**
   * @displayName Active State
   * @validationRule Required for create/update operations
   */
  activeState: boolean;
  /**
   * @displayName Sequence Number
   * @validationRule Required for create/update operations
   */
  sequenceNumber: number;
  /**
   * @displayName Stage Name
   * @validationRule Required for create/update operations
   */
  stageName: string;
  /**
   * @displayName (Deprecated) Stage Type
   * @validationRule Required for create/update operations
   */
  stageTypeKey: RequestStageConfigurationStageTypeKey;
}

export const _RequestStageConfiguration = 'RequestStageConfiguration' as const;