// src/generated/models/request-user-model.ts
import type { Application } from './application-model';
import type { Environment } from './environment-model';
import type { ProvisioningRequest } from './provisioning-request-model';
import type { RequestedUser } from './requested-user-model';

export const RequestUserStatusKeyToLabel = {
  'New': 'New',
  'Provisioned': 'Provisioned',
  'GCCSyncComplete': 'GCC Sync Complete',
  'AccessConfirmed': 'Access Confirmed',
  'NotificationReady': 'Notification Ready',
  'Notified': 'Notified',
  'Exception': 'Exception'
} as const;
export type RequestUserStatusKey = keyof typeof RequestUserStatusKeyToLabel;

export interface RequestUser {
  /**
   * @displayName Request User
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Request User Name
   * @description Human-readable record identifier
   */
  requestUserName: string;
  /**
   * @displayName Application
   */
  application?: Pick<Application, 'id' | 'applicationName'>;
  /**
   * @displayName Environment
   * @validationRule Required for create/update operations
   */
  environment: Pick<Environment, 'id' | 'environmentName'>;
  /**
   * @displayName Exception Reason
   */
  exceptionReason?: string;
  /**
   * @displayName Notes
   */
  notes?: string;
  /**
   * @displayName Provisioning Request
   * @validationRule Required for create/update operations
   */
  provisioningRequest: Pick<ProvisioningRequest, 'id' | 'provisioningRequestName'>;
  /**
   * @displayName Provisioning Ticket Reference
   */
  provisioningTicketReference?: string;
  /**
   * @displayName Requested User
   * @validationRule Required for create/update operations
   */
  requestedUser: Pick<RequestedUser, 'id' | 'fullName'>;
  /**
   * @displayName Status
   */
  statusKey?: RequestUserStatusKey;
}

export const _RequestUser = 'RequestUser' as const;