// src/generated/models/request-user-role-model.ts
import type { AccessRole } from './access-role-model';
import type { RequestUser } from './request-user-model';

export interface RequestUserRole {
  /**
   * @displayName Request User Role
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Request User Role Name
   * @description Human-readable record identifier
   */
  requestUserRoleName: string;
  /**
   * @displayName Access Role
   * @validationRule Required for create/update operations
   */
  accessRole: Pick<AccessRole, 'id' | 'accessRoleName'>;
  /**
   * @displayName Active State
   * @validationRule Required for create/update operations
   */
  activeState: boolean;
  /**
   * @displayName Request User
   * @validationRule Required for create/update operations
   */
  requestUser: Pick<RequestUser, 'id' | 'requestUserName'>;
}

export const _RequestUserRole = 'RequestUserRole' as const;